kodlar karsılastırıldı

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: epaksoy <epaksoy@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/04 11:36:48 by epaksoy           #+#    #+#             */
-/*   Updated: 2023/07/04 12:59:44 by epaksoy          ###   ########.fr       */
+/*   Updated: 2023/07/11 20:29:59 by epaksoy          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,12 @@
 
 size_t	ft_strlen(const char *c)
 {
-	int	x;
+	size_t	x;
 
 	x = 0;
 	while (c[x] != '\0')
 		x++;
 	return (x);
 }
+
+// verilen string'in uzunlugunu donduruyor
